+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## v0.5.0 UNRELEASED
 
+- **breaking TypeScript**: Renamed and removed types. PR #1308
+  - `SxProps` to `SxProp`.
+  - `SxStyleProp`, an alias for `ThemeUIStyleObject` removed. Use `ThemeUIStyleObject` instead.
+- Fix: Add `sx` props types to all props accepting `className`. 
+- Fix WithPoorAsProp to work with ComponentProps utility type.
+
 ## v0.5.0-alpha.2 2020-11-30
 
 - Add Paragraph component. PR #1298
@@ -10,6 +16,7 @@
 
 - Bump React peerDependency to `"^16.14.0 || ^17.0.0"`.
 - Support automatic JSX runtime. Issue #1160, PR #1237
+- Bump React peerDependency to `"^16.14.0 || ^17.0.0"`.
 - Apply variant styles _before_ responsive styles. Issues #1030, and #720, PR #1273
 
 ## v0.5.0-alpha.0 2020-11-20